refactoring, mainly moved some stuff to a new trainers folder
