+++ pps_mw_training/trainers/mlp_trainer.py
@@ -5,23 +5,23 @@ import json
 
 import numpy as np  # type: ignore
 import tensorflow as tf  # type: ignore
-from tensorflow import keras
 from tensorflow.keras.callbacks import ModelCheckpoint  # type: ignore
 from xarray import Dataset  # type: ignore
 
-from pps_mw_training.utils.data import set_missing_data
+from pps_mw_training.models.mlp_model import MlpModel
+from pps_mw_training.utils.augmentation import set_missing_data
 from pps_mw_training.utils.loss_function import quantile_loss
 from pps_mw_training.utils.scaler import Scaler
 
 
 @dataclass
-class QrnnModel:
+class MlpPredictor:
     """
-    Object for handling training, loading, and predictions of a
-    quantile regression neural network model.
+    Object for handling the loading and processing of a quantile
+    regression multi layer perceptron neural network model.
     """
 
-    model: keras.Sequential
+    model: MlpModel
     pre_scaler: Scaler
     post_scaler: Scaler
     input_params: List[str]
@@ -33,20 +33,21 @@ class QrnnModel:
     def load(
         cls,
         model_config_file: Path,
-    ) -> "QrnnModel":
+    ) -> "MlpPredictor":
         """Load model from config file."""
         with open(model_config_file) as config_file:
             config = json.load(config_file)
         input_params = config["input_parameters"]
         output_params = config["output_parameters"]
-        model = cls.create(
+        quantiles = config["quantiles"]
+        model = MlpModel(
             len(input_params),
-            len(output_params),
+            len(output_params) * len(quantiles),
             config["n_hidden_layers"],
             config["n_neurons_per_layer"],
             config["activation"],
-            config["quantiles"],
         )
+        model.summary()
         model.compile()
         model.load_weights(config["model_weights"])
         return cls(
@@ -55,35 +56,69 @@ class QrnnModel:
             post_scaler=Scaler.from_dict(output_params),
             input_params=[p["name"] for p in input_params],
             output_params=[p["name"] for p in output_params],
-            quantiles=config["quantiles"],
+            quantiles=quantiles,
             fill_value=config["fill_value"]
         )
 
     @staticmethod
-    def create(
-        n_input_params: int,
-        n_output_params: int,
-        n_hidden_layers: int,
-        n_neurons_per_layer: int,
-        activation: str,
-        quantiles: List[float],
-    ) -> keras.Sequential:
-        """Create the quantile regression neural network model."""
-        model = keras.Sequential()
-        model.add(keras.Input(shape=(n_input_params,)))
-        for _ in range(n_hidden_layers):
-            model.add(
-                keras.layers.Dense(
-                    n_neurons_per_layer, activation=activation,
-                )
-            )
-        model.add(
-            keras.layers.Dense(
-                n_output_params * len(quantiles), activation="linear"
-            )
+    def prescale(
+        data: Dataset,
+        pre_scaler: Scaler,
+        input_params: List[str],
+    ) -> np.ndarray:
+        """Prescale data."""
+        return pre_scaler.apply(
+            np.column_stack([data[param].values for param in input_params])
         )
-        model.summary()
-        return model
+
+    def postscale(
+        self,
+        data: np.ndarray,
+    ) -> Dataset:
+        """Postscale data."""
+        n = len(self.quantiles)
+        return Dataset(
+            data_vars={
+                param: (
+                    ("t", "quantile"),
+                    self.post_scaler.reverse(
+                        data[:, idx * n:  (idx + 1) * n],
+                        idx=idx,
+                    )
+                ) for idx, param in enumerate(self.output_params)
+            },
+            coords={
+                "quantile": ("quantile", self.quantiles)
+            },
+        )
+
+    def predict(
+        self,
+        input_data: Dataset,
+    ) -> Dataset:
+        """Predict output fron input data."""
+        prescaled = self.prescale(
+            input_data, self.pre_scaler, self.input_params,
+        )
+        prescaled[~np.isfinite(prescaled)] = self.fill_value
+        predicted = self.model(prescaled)
+        return self.postscale(predicted.numpy())
+
+
+@dataclass
+class MlpTrainer(MlpPredictor):
+    """
+    Object for handling training of a quantile regression multi layer
+    perceptron neural network model.
+    """
+
+    model: MlpModel
+    pre_scaler: Scaler
+    post_scaler: Scaler
+    input_params: List[str]
+    output_params: List[str]
+    quantiles: List[float]
+    fill_value: float
 
     @classmethod
     def train(
@@ -108,15 +143,12 @@ class QrnnModel:
         output_path: Path,
     ) -> None:
         """Run the training pipeline for the model."""
-        n_inputs = len(input_parameters)
-        n_outputs = len(output_parameters)
-        model = cls.create(
-            n_inputs,
-            n_outputs,
+        model = MlpModel(
+            len(input_parameters),
+            len(output_parameters) * len(quantiles),
             n_hidden_layers,
             n_neurons_per_layer,
             activation,
-            quantiles,
         )
         learning_rate = tf.keras.optimizers.schedules.CosineDecayRestarts(
             initial_learning_rate=initial_learning_rate,
@@ -128,7 +160,7 @@ class QrnnModel:
         model.compile(
             optimizer=tf.keras.optimizers.Adam(learning_rate=learning_rate),
             loss=lambda y_true, y_pred: quantile_loss(
-                n_outputs, quantiles, y_true, y_pred
+                len(output_parameters), quantiles, y_true, y_pred
             ),
         )
         output_path.mkdir(parents=True, exist_ok=True)
@@ -203,47 +235,3 @@ class QrnnModel:
         ).map(
             lambda x, y: (set_missing_data(x, missing_fraction, fill_value), y)
         )
-
-    @staticmethod
-    def prescale(
-        data: Dataset,
-        pre_scaler: Scaler,
-        input_params: List[str],
-    ) -> np.ndarray:
-        """Prescale data."""
-        return pre_scaler.apply(
-            np.column_stack([data[param].values for param in input_params])
-        )
-
-    def postscale(
-        self,
-        data: np.ndarray,
-    ) -> Dataset:
-        """Transform numpy array holding retrieval data to a dataset."""
-        n = len(self.quantiles)
-        return Dataset(
-            data_vars={
-                param: (
-                    ("t", "quantile"),
-                    self.post_scaler.reverse(
-                        data[:, idx * n:  (idx + 1) * n],
-                        idx=idx,
-                    )
-                ) for idx, param in enumerate(self.output_params)
-            },
-            coords={
-                "quantile": ("quantile", self.quantiles)
-            },
-        )
-
-    def predict(
-        self,
-        input_data: Dataset,
-    ) -> Dataset:
-        """Apply the trained neural network for a retrieval purpose."""
-        prescaled = self.prescale(
-            input_data, self.pre_scaler, self.input_params,
-        )
-        prescaled[~np.isfinite(prescaled)] = self.fill_value
-        predicted = self.model(prescaled)
-        return self.postscale(predicted.numpy())
